backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.39.0
pymupdf==1.24.10
sentence-transformers==3.1.1
psycopg2-binary==2.9.9
pgvector==0.3.4
sqlalchemy==2.0.35
python-dotenv==1.0.1
numpy==1.26.4
gunicorn==22.0.0
